improvements, missed icons complete. dropdowns on edit not perfect

+++ src/app/frame-work/manage/read/fragment/fragment-details/fragment-details.component.ts
@@ -103,6 +103,8 @@ export class FragmentDetailsComponent implements OnInit, AfterViewInit, OnDestro
 
     if (!this.fragmentManagerService.verificationDetails(dataSource['dataSource']))
       return;
+    const confirmed = await this.fragmentManagerService.firstConfirmDialog();
+    if (!confirmed) return;
     const a = await this.fragmentManagerService.removeFragmentDetails(dataSource['dataSource']);
     if (a) {
       this.dataSource[dataSource['ri']] = this.clonedProducts[dataSource['dataSource'].fragmentMasterId];
